(PC-19593) refactor(navigation): move useGoBack tests outside __tests__ folder
